Load Balancer Issue (#30)
For ssl websites behind load balancers (e.g. Amazon's AWS) or reverse proxies that support HTTP_X_FORWARDED_PROTO this fixes an issue of not being recognized as using https.  It should be noted that Wordpress devs makes mention of this in their docs saying that sites in this situation can simply add a similar line to their wp-config.php file, but i figure most people won't realize this and just assume its a w3tc bug.  Better to just have it built in.