Fix for Role Removal from $current_user (#49)
When the $current_user is logged in (meaning, they have a user id > 0)
they're always guaranteed to have a designated "role".  With that in
mind one can do away with the ill advised array_shift() function call,
which was unintentionally stripping the active $current_user's role, and
simply just lookup the 1st item in its "roles" array.