Amazon Cloudfront - Expires header (#66)
It seems w3tc intentionally does not permit the use of the Expires
header when "Amazon Cloudfront" (origin pull) is selected as a CDN.  The
following link shows Amazon does support the Expires header and so this
commit re-enables the ability to use Expires:
http://docs.aws.amazon.com/AmazonCloudFront/latest/DeveloperGuide/Expiration.html#ExpirationDownloadDist

- A bug was found whereby despite having CDN disabled, if you select
"Amazon Cloudfront" (origin pull) all Expires headers checkboxes (under
Browser Cache) become disabled and unchecked.  The fix would have been
to do a verify that the CDN checkbox was enabled but because the Expires
header is now supported by Amazon i just ended up removing the disabling
ability entirely.