Sitemap default priority change
Although plugins like Yoast have dropped the use of the priority
attribute w3tc required that the priority existed for each url entry in
a sitemap.  If such priority did not exist (despite a correlating url)
it would ignore the entry entirely, and features like prime caching
would no longer work for all entries that did not have a priority, which
in most cases would be the entire sitemap.

By simply modifying the default priority to 0.5 (the standard default
value for this attribute) it corrects this problem and allows urls with
no priority  to continue to be processed.