Fixes Feed Caching Issue and Content-Type for Feeds and Sitemaps
This allows for proper caching and client-side returns of feeds and sitemaps for disk enhanced when the "Handle XML Mime Type" checkbox is selected. 

Other notables:  Minifying of feeds and sitemaps also worked when verified using W3C Feed Validator tool.  A minor tweak was made to the Purge Sitemaps default regex.  For fresh installs i set it so sitemaps aren't cached by default.  A php warning notice was fixed in Minify.php that occurred when attempting to minify an xml file. It was looking for a head tag when none exist in xml files.