Hide Update notification & Added Fragment Caching Exception Handling  (#127)
1. Adds a new "Hide Plugin Update Notification" checkbox under General
Settings' Miscellaneous area. This allows the user to decide if they
want to hide the WordPress notification message about updating to the
latest w3tc.

2. This resolves a minor issue when users implement <--mfunc but trigger
a ParseError exception when using php7. Prior to php7 the eval()
function returned false when an error occurred. w3tc would then display
a friendly error message on screen and resume. As of php7 the code is
halted on parse errors. This revision makes it so that when using php7 a
parsing error will act like previous php versions when using w3tc by
informing the user but continuing to process the page.