WP-CLI Update
- New ability to now prime from wp-cli --> wp w3-total-cache prime
<ARGS>

- An improvement was made to the unschedule_prime() function.  If a
prime is set up to only do x amount of pages per interval and then the
prime is stopped via unchecking the "Automatic Cache Preload", you could
see the prime continues to chug along until the entire sitemap has been
exhausted.  This occurred because the prime args were not passed to
wp_unschedule_event() causing it to not find the active event.  Because
the args are dynamic the fix was not as straight forward.