Gzip Cached Pages are not Decoding (#313)
This fix resolves a critical issue whereby headers are being sent back incorrectly.  This has a dramatic effect when gzip'ed cached pages are sent back to the client browser -- they are unable to be decoded and thus, pages show up all garbled.

This only occurs in php 5.3 and lower.  What is happening is php 5.3's (and lower) isset() function interprets the parameter given (ie, $value0['n']) as a string of characters not a placeholder for an array and so the isset() ends up being TRUE. And the index value of 'n' ends up being index 0 into this string. What that does is it returns the first letter of the header's value back to the user e.g. "Cache-Control: max-age=3600, public" becomes "m: m". This peculiar behavior is not present in php 5.4 and higher.