CSSTidy Update (#317)
This is an update to the aging CSSTidy engine. Bumps it up to v1.5.5 (the latest) which includes support for CSS3.0. I also added bug fixes to this engine. And also includes a few new user-selectable options under the Minify > CSS section.

This new version of CSSTidy requires at least php 5.4 and so will disable itself (within the General Settings > Minify > CSS minifier drop-down box) if that requirement is not met.