Page Cache - New Fields (#319)
Found under Page Cache's Advanced section, i created 4 new user boxes:

* Never Cache Pages Associated With These Categories
* Never Cache Pages That Use These Tags
* Never Cache Pages By These Authors
* Never Cache Pages That Use These Custom Fields

This will help give more granular control over groups of related pages -- whether they should be cached or not.  It was a feature request by a few in the WP support forum.

For example, if someone wanted to not cache all posts that appear under a specific category (e.g. logged in visitors have a "Members" section) then prior to this update each post would have to be manually added into the "Never cache the following pages" box. With this update one can now, alternatively, provide the category slug name within the "Never cache pages associated with these categories" box and all existing and future posts filed under this category will not be cached. This saves a lot of time and potential mistakes (i.e, not remembering to update the "Never cache the following pages" box).

In the new "Never Cache pages that use these custom fields" box, a default value of "w3tc=dontcache" is given. Logged in users who create posts can now decide if they don't want their posts cached. They simply add this custom field (under the "Custom Fields" section of a post) and that's it. For sites that have multiple users who can write posts this will give them better control without needing to contact the admin repeatedly to configure w3tc to not cache their posts.

I also fixed a bug for the Cache Exception List. Prior to this update, while analyzing the code i noticed that the design made it very unlikely for entries in this list to work. And also, despite what this field's description claimed about superseding matched entries found in the "Never Cache" fields, that was not the case. It works now.