WP-CLI - Prime the Page Cache (#324)
This adds the much desired cache preload feature to W3 Total Cache's WP-CLI space. It's the ability to prime the page cache from the command line. It was a feature request by WP forum users to ease the burden of script-managed re-caching of multiple sites remotely. It was also added to help flesh out W3 Total Cache's completeness when using WP-CLI.

Because the priming can potentially take very long to process it has been designed to run asynchronously in the background after initiating the request, thus, freeing up the command line prompt immediately. There is also a command to stop the running prime process, if needed.

This also fixes a few minor bugs in Cli.php: The Exception object was being incorrectly passed to the WP_CLI::Error() function in some areas, ironically, causing an exception error of its own. It seems the sprintf() wrapper was missing in a few spots and the object's getMessage() call was also missing. To save on an extra function call i just opted to concatenate the exception message string into their proper positions.