Amazon Web Services (AWS) v4 Support & New Locations (#332)
Because Amazon has moved AWS to signature version 4 (v4) it left W3TC incapable of authenticating AWS requests to v4 regions and endpoints. At the time of this writing there are six "Version 4 Only" regions. And Amazon has made it clear that all new regions and endpoints created in the future will only be for v4.

Prior to this update when one made a request to "v4 only" regions in W3TC you got an AWS4-HMAC-SHA256 error message. AWS4-HMAC-SHA256 is the signing algorithm for the v4 signature. This would have been a problem if the user had created a bucket outside W3TC (e.g., on the AWS website) that used a "v4 only" region and provided the bucket name in W3TC.

With this update the underlying CDN engine now supports signature version 4. It also maintains backwards compatibility to all version 2 regions so the user does not need to make any changes -- it will use v4 when it needs to.

Interestingly, I noticed that W3TC only made available four selectable regions when creating a bucket. I assume it was because this list had not been updated for a very long time. And so, in this update I've also gone ahead and revised this list to include all available regions (14 in total, as of this writing).

Design Layout Note:

I had to slightly modify how the "Create Bucket" button (of "Amazon Simple Storage Service: S3 - Origin Push") and the "Create Bucket & Distribution" button (of "Amazon CloudFront - Origin Push") operated. Prior to this update these buttons would bring up a popup modal window allowing the user to select a bucket location/region before creating the bucket. After the bucket was generated the selected region associated with that bucket was never shown again to the user. That was fine for version 2 regions but with v4 a new problem emerged.

Because the v4 signature demands that an existing bucket name also include its associating bucket region for all AWS requests it is now imperative that the region always be shown on the same page as the bucket name so that those who create a "v4 only" bucket outside W3TC can now provide their selected region within W3TC. If they aren't able to provide that associated region, as was the case previously, all communication will fail. As such, the region drop-down list box was moved and placed alongside the bucket name.

Current W3TC users of AWS, prior to this update, will not be affected by this change in any way. The drop-down box will simply show the words "Select Bucket Location", while under the hood it will be using the version 2 signature for all AWS communications -- just like before.