Save Cloudflare Settings - SSL Update Failure (#334)
After clicking the "Save Cloudflare Settings" button within the CloudFlare settings' extensions page a "Failed to update option ssl: Invalid value for zone setting, ssl" alert message is seen. This error seems to occur consistently when the SSL drop-down box is modified to Off.

Why The Problem Occurs

This error occurs because the words: on and off are special indicators to W3TC when dealing with checkbox values. Because the SSL drop-down list box is not a checkbox and in fact uses the keyword off, W3TC was mistakenly confusing it as a checkbox and, thus, setting the final SSL value as the word: on.

But why on? Well, it does this because since it mistakenly assumed it was dealing with a checkbox it figured the value had to be an integer. But since the value: off, which the SSL drop-down sent to W3TC, is not an integer it defaulted to setting the final value, which is sent to Cloudflare, as the word: on. Now, because Cloudflare doesn't understand the word on, under the context for setting the SSL, it returned back a failure response.