Malformed HTML in Admin Pages (#343)
When using the Util_Ui::config_item function and setting it's style attribute to 2 (this tells it to span two columns for its container table header), it causes the HTML to be malformed, resulting in the generated <label> tag to be inserted incorrectly, potentially destroying the page layout in less popular (or older) browsers, which might even include Internet Explorer. This fix resolves the problem and the page validates correctly.