Save Settings Strips Backslash (#361)
This was a bug that stripped out a single backslash for every pair in user input prior to saving the settings. This would typically be a problem if using it in regex. So, for example, something like "sitemap\.xml" would have been saved as "sitemap.xml", while "sitemap\\.xml" would have been saved as "sitemap\.xml"

The problem was due to a previous optimization. The optimization dramatically reduces the amount of time that php has to generate a new array. What wasn't known until now is that some function earlier in the chain modifies the $_POST content which adds double-slashes prior to saving the settings. It is triggered by the init action. Since the optimization was unaware of this it saved the user data without the modification, thus, appearing to strip out the backslashes.

This fix resolves the problem.