Adds Helpful Default Values to CSP (Security Headers) & Grammar Corrections (#363)
This is an amendment to #344 (the new "Security Headers" section to W3TC).

Because the Content Security Policy (CSP) section can seem daunting at first for inexperienced users I felt it is important to include default values and placeholder field examples that users can rely on and use for this security feature. It makes for a better experience when using W3TC.

This amendment also fixes some grammatical errors I discovered in the Security Headers section. Sorry.