Fix for the "Trying to Get Property of Non-Object" Message (#376)
Sometime the error message: "Trying to Get Property of Non-Object" appears in the error log, but the cause is unknown.  A hypothesis is that the garage collector tried to purge a deleted post from an old sitemap.

In accordance with the WordPress documentation, get_post() returns WP_Post object type or null (in the context of how w3tc is using it).  Because no check was done for null this left open the door for the error message shown above.  This fixes the problem.