Browser Cache - Keeping WP Query String (#384)
When "Browser Cache" is enabled and the user makes use of the "Prevent Caching of Objects After Settings Change" feature, W3TC postfixes a unique query string identifier to an object's URL request. It was discovered that the WP query string name-value pair (i.e., "?ver=4.7.2"), which is typically added at the end of these URLs, was being stripped. This only happens when "Prevent Caching of Objects After Settings Change" is enabled. In other words, when that feature is disabled the WP query string is never stripped, and only gets stripped when using the W3TC "Minify" feature. This seems to indicate it is a bug.

There are legitimate cases for not stripping the WP query string under this context.  For example,  a user relied on setting the version query value of scripts and styles to their respective last modified dates to better control caching.