Malformed WP Query String (#386)
Very similar bug to # 384 but for the v0.9.4.x generation. There are differences though in that the bug in v0.9.5.x happened all the time when the "Prevent Caching of Objects After Settings Change" feature was enabled. This bug for v0.9.4.x only occurred when "Prevent Caching of Objects After Settings Change" was enabled and CDN was disabled.

However, when CDN was enabled a uniquely different bug occurred. It would result in malformed URLs having two "?", one for the unique identifier set by W3TC, and another for the "?ver" WP query string identifier.