Feeds Not Caching Nor Serving Back as XML (#393)
When using the "Cache Feeds" checkbox under "Page Cache" it is suppose to cache feeds as identified by WP's "is"feed()" function.  However, W3TC assumed the content type of feeds was just "text/xml" (which was true in older WP versions) but current WP iterations (for some time now) set the feed content types to "application/rss+xml", "application/atom+xml", etc.  As such, prior to this fix W3TC did not cache any feeds (that i know of).

With this patch, when for example "Disk: Enhanced" is used, you will see a directory called "feed" created containing the  cached file of the feed.

Several tests show the problem is fixed and works really well.  I can cache feeds (e.g., rss, rss2, atom, rdf) and retrieve them from cache properly.  Because caching feeds finally are working I recommend people use the "Purge Policy" checkboxes to manage when these cached feeds get cleared (e.g., when its associating post is updated or a new comment is made).

There was also a bug that preventing feeds and sitemaps from minifying.  It's now fixed and they validate successfully when using W3C's Feed Validator tool.

Interestingly, it was discovered when using "Disk: Enhanced" W3TC always stores feeds and xml sitemaps as "index.html", regardless of content type.  Because w3tc does not have any way of identifying what that content type was after being stored as "index.html" it always serves back all pages as content type: "text/html".  In other words, it doesn't set the "Content-Type:" header for "Disk: Enhanced" responses and so the server is left to be the final decider by looking at the file extension.  As such, xml sitemaps which are suppose to return back as Content-Type: "text/xml" or rdf feeds as "application/rdf+xml", will both return as "text/html" to the client.  This will render client-side results visually incorrect (in some cases: garbled) even though the contents of that page will be proper.  So this is not a problem with the content it is a problem with the returning mime type being set (or more correctly, not being set) for "Disk: Enhanced".

For NGINX users this problem is solved by using the "Handle XML Mime Type" checkbox, which appears at the bottom of "Page Cache", allowing W3TC to set feeds and xml sitemap cached files using a ".xml" file extension.  This patch extends this ability by offering the same checkbox for Apache users as well.  It even goes one step further for "Disk: Enhanced" mode by updating the ".htaccess" file to be aware of potentially cached xml types -- improving performance.

So, in summation this fix allows feeds and xml sitemaps to be cached and minified correctly and return back their proper mime content-type so they can render accurately.