Catch Exceptions Thrown When Saving Config (#408)
For environments that do not allow the configuration file to be written, as of 0.9.5.2 this causes fatal errors on any W3TC settings page and on the plugins page in the admin interface.  The resulting error message is: "PHP Fatal error:  Uncaught exception 'Exception' with message 'Can't write to file /w3tc-config/master.php".

By adding a try/catch block to catch this exception the problem is now fixed.