Fix a timeout update that was missed in push CDN upload PR (#525)
An earlier PR, https://github.com/szepeviktor/w3-total-cache-fixed/pull/452,
Fixed several bugs in the code for uploading files to push CDNs,
including increasing timeouts for uploading chunks of files, since
five seconds wasn't nearly enough in many cases. Unfortunately, one of
the timeouts that needed to be increased was missed. This fixes that
missed timeout so it matches the others.