Write correct data into minify cache file when cache dir is created (#524)
There was a bug in the code for detecting when the minify cache
directory does not exist while trying to save to a file in it,
creating it and attempting to save the file again. In this situation,
the wrong data would be saved into the cache file. This commit fixes
this problem.